requirements.txt
requests==2.28.1
Click==8.1.3
numpy~=1.23.3
Pillow~=9.0.1
scipy>=0.17.0
pathlib~=1.0.1
PyYAML~=6.0
pytest~=3.6.3

# face recognition
face-recognition==1.3.0
face_recognition_models==0.3.0
deepface~=0.0.75
dlib>=19.3.0

# flask
Flask==2.2.2
py-healthcheck==1.10.1
